Refactor TextEdit with partial classes and focus fix
